+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to `gosms` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.1.3] - 2020-06-12
+
+### Added
+- Add ResponseResolverInterface for create custom response resolver
+- Add new value object SentMessage
+- Add toArray methods for value objects
+
+### Changed
+- Modify ResponseResolverObject resolver
+- Modify GoSMS for easier implementation
+
+## [0.1.2] - 2020-06-11
+
+### Fixed
+- Fixed name of variable, when you have not access token
+
 ## [0.1.1] - 2020-06-10
 
 ### Changed